% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessor.R
\name{addDeptName}
\alias{addDeptName}
\title{Functions for processing raw NorSpis data}
\usage{
addDeptName(df, id, name)
}
\arguments{
\item{df}{Data frame of registry data}

\item{id}{Vector of unique ids for organization names. Must be of the same
length as \code{name}. Normally this should be an ReshId.}

\item{name}{Vector of unique organization names for corresponding ids. Must
be of the same length as \code{id}.}
}
\value{
A data frame of registry data with a new column providing department
or organizational name based on AvdRESH
}
\description{
Functions for processing raw NorSpis data
}
\examples{
# sample data
df <- data.frame(AvdRESH = c(1, 2, 3))
id <- c(1, 2, 3)
name <- c("A", "B", "C")

addDeptName(df, code, value)
}
