% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab}
\alias{admtab}
\alias{admtab_ui}
\alias{admtab_server}
\alias{admtab_demo}
\title{Shiny modul for administrative tabeller for Norspis}
\usage{
admtab_ui(id)

admtab_server(id, skjemaoversikt)

admtab_demo()
}
\arguments{
\item{id}{Tekststreng med shiny modul id.}

\item{skjemaoversikt}{dataramme med oversikt over registreringer i Norspis}
}
\value{
Shiny objekter for Norspis sine administrative tabeller
}
\description{
Kun til bruk i Shiny, returner modulfunksjoner
}
