% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump}
\alias{datadump}
\alias{datadump_UI}
\alias{datadump_ui}
\alias{datadump_server}
\title{Modul for datadump-fane i Norspis sin shiny-app på Rapporteket}
\usage{
datadump_UI(id)

datadump_server(id, RegData, userRole, hvd_session, reshID)
}
\description{
Kun til bruk i Shiny
}
