% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deptNames.R
\docType{data}
\name{deptNames}
\alias{deptNames}
\title{Resh ids and department names for NorSpis}
\format{
A data frame with 2 variables:
\describe{
\item{reshId}{Organization id for a department}
\item{deptName}{Organization or department name}
...
}
}
\usage{
deptNames
}
\description{
A dataset containing the Resh ids and corresponding department names for
NorSpis.
}
\keyword{datasets}
