% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisRegDataSQL.R
\name{NIRRegDataSQL}
\alias{NIRRegDataSQL}
\title{Henter data registrert for NorSpis}
\usage{
NIRRegDataSQL(datoFra = "2017-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}
}
\value{
Lager dataramma RegData for utvalgte data fra NorSpis
}
\description{
Henter data for NorSpisregisteret fra "staging"
}

