% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norspisBeregnIndikator.R
\name{norspisBeregnIndikator}
\alias{norspisBeregnIndikator}
\title{Beregn kvalitetsindikatorer for Norspis}
\usage{
norspisBeregnIndikator(RegData, ind_id = "norspis_KI1_symptomreduksjon_EDEQ_V")
}
\description{
Beregn kvalitetsindikatorer for Norspis til bruk på Rapportalen
}
