% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisELAlleScorData.R
\name{NorSpisELAlleScorData}
\alias{NorSpisELAlleScorData}
\title{Hente data fra enkeltspørsmål (enkeltledd), NorSpis}
\usage{
NorSpisELAlleScorData(datoFra = "2015-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Hoveddato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Hoveddato, til og med. Standard: '3000-01-01' (siste registreringsdato)}
}
\value{
RegData data frame
}
\description{
Henter data fra staging
}

