% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisFigAndeler.R
\name{NorSpisFigAndeler}
\alias{NorSpisFigAndeler}
\title{Søylediagrammet viser andelsfordelinga til den valgte variabelen. Søylene er horisontale eller vertikale
avhengig av hvor stor plass kategorinavnet til søyla tar.}
\usage{
NorSpisFigAndeler(RegData, valgtVar, datoFra = "2016-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  outfile = "", hentData = 0, preprosess = 1, reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Hoveddato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Hoveddato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{outfile}{Navn på fil figuren skrives ned til. outfile lik '' skriver til skjerm.}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}
}
\value{
En figur med søylediagram (fordeling) av ønsket variabel
}
\description{
#' Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling, 10-årige grupper
    \item B02EgneBarn
    \item B03Bosituasjon
    \item B04PabegyntUtd
    \item B05FullfortUtd
    \item B06Hovedaktivitet
    \item B07Hovedinntekt
    \item B08StartAldrProbl
    \item B11FamilieSF
    \item B12TidlBehSF
    \item B12cAldrForsteBeh
    \item B17FysMishandl
    \item B18PsykMishandl
    \item B19Overgrep
    \item B20Mobbing
    \item B21SelvskadTidl
    \item B22SelvskadSisteAr
    \item B23SelvmordFTidl
    \item B24SelvmordFSisteAr
    \item B25Avhengighet
    \item BehUtfallsvurdSamlet
    \item BehVidereBeh
    \item HCA01Atferd
    \item HCA02Aktivitetsniva
    \item HCA03Selvskade
    \item HCA04Rusmisbruk
    \item HCA05SkoleSprak
    \item HCA06FysiskProblem
    \item HCA07Hallusinasjoner
    \item HCA08SomatiskSymp
    \item HCA09EmosjonelleSymp
    \item HCA10JevnaldrProbl 
    \item HCA11Egenomsorg
    \item HCA12FamilieProbl
    \item HCASkoleframmote
    \item HCA14ProblKunnskap
    \item HCA15Mangelinfo
    \item H01Atferd
    \item H02Selvskade
    \item H03Rusmisbruk
    \item H04KognitiveProbl
    \item H05FysiskeProbl
    \item H06Hallusinasjoner
    \item H07Stemningsleie
    \item H08AndreProbl
    \item H09ForhAndre
    \item H10ADLProbl
    \item H11BoligProbl
    \item H12YrkeProbl
    \item MedAntidepressiva
    \item MedBenzodiazepiner
    \item MedBMISlutt: Pasientenes BMI ved slutt av behandling
    \item MedBMIStart: Pasientenes BMI ved innkomst til behandling
    \item MedIsoBMIBGSSlutt: Pasientenes iso-BMI (basert på BGS-normdata) ved slutten av behandlingen
    \item MedIsoBMIBGSStart: Pasientenes iso-BMI (basert på BGS-normdata) ved innkomst 
    \item MedIsoBMICDCSlutt: Pasientenes iso-BMI (basert på CDC-normdata) ved slutten av behandlingen
    \item MedIsoBMICDCStart: Pasientenes iso-BMI (basert på CDC-normdata) ved innkomst 
    \item MedNevroleptika
    \item NegHend
    \item Norsktalende: Snakker pasienten norsk
    \item PT01OnsketInvolv
    \item PT02BleInvolv
    \item PT03Utfallsvurd
    \item PT04KontaktBrukerorg
    \item PT05OrientertBrukerorg
    \item RegHenvInstans
    \item TidSykBehandling
    \item VentetidKat: Ventetid, 2-ukers grupper 
   }
}
