% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisFigAndelerGrVar.R
\name{NorSpisFigAndelerGrVar}
\alias{NorSpisFigAndelerGrVar}
\title{Søylediagram med AggVerdier for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NorSpisFigAndelerGrVar(RegData, valgtVar, datoFra = 0, datoTil = 0,
  minald = 0, maxald = 130, grVar = "", erMann = "", hentData = 0,
  preprosess = 1, outfile = "", lagFig = 1)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Hoveddato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Hoveddato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{outfile}{Navn på fil figuren skrives ned til. outfile lik '' skriver til skjerm.}
}
\value{
Søylediagram med AggVerdier av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med AggVerdier av en gitt variabel for ei valgt gruppering, 
f.eks. enheter. I øyeblikket benytter funksjonen bare 'EnhNavn' som grupperingsvariabel, men 
andre valg kan lett inkluderes.
}
\details{
Figurtypen som genereres er avhengig av valgtVar. Ved valgtVar='InnMaate', får man ei oversikt oversikt
over andel av innkomsttype ved hver enhet, dvs. ei fordeling av de tre innkomsttypene for hver enhet.
For andre "valgtVar" viser figuren andel av den valgte variabelen for hver enhet.
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder_u18: Pasienter under 18 år
    \item BehDodUnderBeh: Andel pasienter som døde i behandling
    \item DiagVDiabetes: Andel pasienter med diabetes
    \item VentetidOver2Uker: Andel med ventetid over 2 uker, beregnet fra dato for mottatt henvisning til hendelsesdato.
   }
Funksjonen benytter funksjonene: NorSpisRegDataSQL, NorSpisPreprosess, NorSpisVarTilrettelegg, NorSpisUtvalg
og NIRFigSoyler
}
