% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norspisLesOgProsesser.R
\name{norspisLesOgProsesser}
\alias{norspisLesOgProsesser}
\title{Les og preprosesser data for Norspis}
\usage{
norspisLesOgProsesser()
}
\description{
Leg nødvendige tabeller for Norspis og returner en utflatet dataramme
}
