% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisPreprossesser.R
\name{NorSpisPreprosess}
\alias{NorSpisPreprosess}
\title{Preprosesser data fra NorSpis}
\usage{
NorSpisPreprosess(RegData = RegData)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}
}
\value{
Data En liste med det filtrerte datasettet (og sykehusnavnet som tilsvarer reshID, ikke pt)
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
}
