% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisRegDataSQLikkeIbruk.R
\name{NorSpisRegDataSQL}
\alias{NorSpisRegDataSQL}
\title{Henter data registrert for NorSpis}
\usage{
NorSpisRegDataSQL(datoFra = "2015-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Hoveddato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Hoveddato, til og med. Standard: '3000-01-01' (siste registreringsdato)}
}
\value{
Lager dataramma RegData for utvalgte data fra NorSpis
}
\description{
Henter data for NorSpisregisteret fra "staging"
}
