% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryData.R
\name{queryData}
\alias{queryData}
\alias{queryAlleScorer}
\alias{queryBehandling}
\alias{queryBehandlingNum}
\alias{queryEnkeltLedd}
\alias{queryEnkeltLeddNum}
\alias{queryForlopsOversikt}
\alias{queryReshNames}
\title{Get data from NorSpis}
\usage{
queryAlleScorer(registryName, reshId, ...)

queryBehandling(registryName, reshId, ...)

queryBehandlingNum(registryName, reshId, ...)

queryEnkeltLedd(registryName, reshId, ...)

queryEnkeltLeddNum(registryName, reshId, ...)

queryForlopsOversikt(registryName, reshId, ...)

queryReshNames(registryName, ...)
}
\arguments{
\item{registryName}{Character string with name of the registry from which
data are to be queried}

\item{reshId}{Integer providing the organization id. Usefull for data
filtering}

\item{...}{Optional arguments to pass to the function. If \code{session} is
provided this will be assumed a valid shiny session object and hence logging
will be performed}
}
\value{
Data frame of registry data
}
\description{
Functions to query data from a database holding NorSpis data at Rapporteket.
Providing these function with the shiny session object logging of queries
will also be performed
}
