% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProcessor.R
\name{reportProcessor}
\alias{reportProcessor}
\title{Common report processor for NorSpis}
\usage{
reportProcessor(
  report,
  outputType = "pdf",
  title = "",
  author = "unknown author",
  orgName = "unknown organization",
  orgId = 999999,
  registryName = "norspis",
  userFullName = "unknown person name",
  userRole = "unknown role"
)
}
\arguments{
\item{report}{Character string identifying the report to be processed by this
function.}

\item{outputType}{Character string with output format. Must be one of
\code{c("html", "pdf")}. Defaults to "pdf".}

\item{title}{Character string giving the report title. Empty string by
default.}

\item{author}{Character string providing the report author. Default value is
"unknown author".}

\item{orgName}{Character string with the name of the organization/hospital.
Default is "unknown organization".}

\item{orgId}{Integer (?) with the id of the organization/hospital. Default is
999999.}

\item{registryName}{Character string with registry name. Default is
"norspis".}

\item{userFullName}{Character string giving the person name, normally the
user requesting the report. Default is "unknown person name".}

\item{userRole}{Character string giving a user role, normally the one of the
user requesting the report. Default is "unknown role".}
}
\value{
A character string with a path to where the produced file is located.
}
\description{
Makes reports for NorSpis typically used for auto reports such as
subscriptions, dispatchments and bulletins. As such, please be warned that
any changes to this function might render existing auto reports nonfunctional
as they are based on static calls based on any previous version of this
function. Changes should therefore be thoroughly tested against existing auto
reports. Altering the names of the arguments will likely be a breaking
change. Adding new arguments should be safe as long as they are provided a
default value.
}
\examples{
## Make the start page for NorSpis
reportFilePath <- reportProcessor(report = "veiledning",
                                  outputType = "html",
                                  title = "Example report")
}
