% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stottefunksjoner.R
\name{separer}
\alias{separer}
\title{Trekk ut diagnoser fra kolonne med kommaseparerte diagnoser til separate
kolonner}
\usage{
separer(variable, separator = ",", unique_varnames = NULL)
}
\arguments{
\item{variable}{\itemize{
\item kolonnen som skal splittes i flere
}}

\item{separator}{\itemize{
\item separator mellom elementer i tekststreng. Overflødig hvis
unique_varnames gis som input
}}

\item{unique_varnames}{\itemize{
\item vektor med navnene som skal trekkes ut i nye kolonner, må/bør
samsvare med navn som eksisterer i kolonnen som skal splittes
}}
}
\value{
En dataramme med separerte variabler
}
\description{
Trekk ut diagnoser fra kolonne med kommaseparerte diagnoser til separate
kolonner
}
