% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stottefunksjoner.R
\name{tr_summarize_output}
\alias{tr_summarize_output}
\title{Transponer output fra dplyr::summarise}
\usage{
tr_summarize_output(x, kolnavn1 = "")
}
\value{
tr_frame Den transponerte av inputen
}
\description{
Denne funksjonen tar som input resultatet av dplyr::summarise og returnerer dens
transponerte uten at formatene endres.
}
\details{
Her kan detaljer skrives
}
