% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisFigAndeler.R
\name{NorSpisFigAndeler}
\alias{NorSpisFigAndeler}
\title{Søylediagram, horisontalt eller vertikalt, som viser andeler av valgt variabel.}
\usage{
NorSpisFigAndeler(RegData, valgtVar, datoFra = "2016-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  outfile = "", hentData = 0, preprosess = 1, reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.}

\item{datoFra}{Hoveddato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Hoveddato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{outfile}{Navn på fil figuren skrives ned til. outfile lik '' skriver til skjerm.}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}
}
\value{
En figur med søylediagram (fordeling) av ønsket variabel
}
\description{
Søylediagrammet viser fordelinga til den valgte variabelen. Søylene er horisontale eller vertikale
avhengig av hvor stor plass kategorinavnet til søyla tar.
}
\details{
#' Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Aldersfordeling, 5-årige grupper
    \item BMI: Pasientenes BMI (Body Mass Index)
    \item Norsktalende: Snakker pasienten norsk
    \item Utdanning: Utdanningsnivå
    \item 
   }
}
