% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpisUtvalg.R
\name{NorSpisUtvalg}
\alias{NorSpisUtvalg}
\title{Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
NorSpisUtvalg(RegData, datoFra, datoTil, minald = 0, maxald = 130,
  erMann = "", InnMaate = "", fargepalett = "BlaaOff")
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Standard: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.
}

